# src/main/java/atlassports/api/AuthApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package atlassports.api;

import atlassports.model.dto.AuthRequest;
import atlassports.model.dto.UpsertUserDto;
import atlassports.model.dto.UserDto;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import javax.annotation.processing.Generated;
import java.util.Optional;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-24T16:37:25.216838800+01:00[Europe/Skopje]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "Auth", description = "Authentication and authorization")
public interface AuthApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_AUTH_LOGIN_POST = "/auth/login";

    /**
     * POST /auth/login : Login user
     *
     * @param body (required)
     * @return Successful login (status code 200)
     */
    @Operation(
            operationId = "authLoginPost",
            summary = "Login user",
            tags = {"Auth"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "Successful login", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = AuthApi.PATH_AUTH_LOGIN_POST,
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    default ResponseEntity<String> authLoginPost(
            @Parameter(name = "body", description = "", required = true) @Valid @RequestBody AuthRequest body
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_AUTH_REGISTER_POST = "/auth/register";

    /**
     * POST /auth/register : Register user
     *
     * @param body (required)
     * @return Successful register (status code 200)
     */
    @Operation(
            operationId = "authRegisterPost",
            summary = "Register user",
            tags = {"Auth"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "Successful register", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = AuthApi.PATH_AUTH_REGISTER_POST,
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    default ResponseEntity<UserDto> authRegisterPost(
            @Parameter(name = "body", description = "", required = true) @Valid @RequestBody UpsertUserDto body
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}