# src/main/java/atlassports/api/BookingsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package atlassports.api;

import atlassports.model.dto.BookingDto;
import atlassports.model.dto.UpsertBookingDto;
import atlassports.model.openapi.BookingsMeUpcomingGet200Response;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import javax.annotation.processing.Generated;
import java.util.Optional;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-24T16:37:25.216838800+01:00[Europe/Skopje]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "Bookings", description = "Booking operations")
public interface BookingsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_BOOKINGS_GET = "/bookings";

    /**
     * GET /bookings : Get all bookings (Admin can filter by tenant)
     *
     * @return List of bookings (status code 200)
     */
    @Operation(
            operationId = "bookingsGet",
            summary = "Get all bookings (Admin can filter by tenant)",
            tags = {"Bookings"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "List of bookings", content = {
                            @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Object.class)))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = BookingsApi.PATH_BOOKINGS_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<Page<BookingDto>> bookingsGet(
            Pageable page
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ null, null ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_BOOKINGS_ME_HISTORY_GET = "/bookings/me/history";

    /**
     * GET /bookings/me/history : Get past bookings for current user
     *
     * @param page (optional, default to 0)
     * @param size (optional, default to 20)
     * @param sort (optional)
     * @return Paginated list of past bookings (status code 200)
     */
    @Operation(
            operationId = "bookingsMeHistoryGet",
            summary = "Get past bookings for current user",
            tags = {"Bookings"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "Paginated list of past bookings", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = BookingsMeUpcomingGet200Response.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = BookingsApi.PATH_BOOKINGS_ME_HISTORY_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<Page<BookingDto>> bookingsMeHistoryGet(
            Pageable page
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"size\" : 6, \"last\" : true, \"totalPages\" : 5, \"page\" : 0, \"content\" : [ null, null ], \"first\" : true, \"totalElements\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_BOOKINGS_ME_UPCOMING_GET = "/bookings/me/upcoming";

    /**
     * GET /bookings/me/upcoming : Get upcoming bookings for current user
     *
     * @param page (optional, default to 0)
     * @param size (optional, default to 20)
     * @param sort (optional)
     * @return List of upcoming bookings (status code 200)
     */
    @Operation(
            operationId = "bookingsMeUpcomingGet",
            summary = "Get upcoming bookings for current user",
            tags = {"Bookings"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "List of upcoming bookings", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = BookingsMeUpcomingGet200Response.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = BookingsApi.PATH_BOOKINGS_ME_UPCOMING_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<Page<BookingDto>> bookingsMeUpcomingGet(
            Pageable page
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"size\" : 6, \"last\" : true, \"totalPages\" : 5, \"page\" : 0, \"content\" : [ null, null ], \"first\" : true, \"totalElements\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_BOOKINGS_POST = "/bookings";

    /**
     * POST /bookings : Create a booking
     *
     * @param body (required)
     * @return Booking created (status code 201)
     */
    @Operation(
            operationId = "bookingsPost",
            summary = "Create a booking",
            tags = {"Bookings"},
            responses = {
                    @ApiResponse(responseCode = "201", description = "Booking created", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = BookingsApi.PATH_BOOKINGS_POST,
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    default ResponseEntity<BookingDto> bookingsPost(
            @RequestBody UpsertBookingDto upsertBookingDto
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /bookings/tenants/{tenantId} : Get all bookings for tenant
     *
     * @param tenantId (required)
     * @return List of bookings (status code 200)
     */
    @Operation(
            operationId = "bookingsTenantsTenantIdGet",
            summary = "Get all bookings for tenant",
            tags = {"Bookings"},
            responses = {
                    @ApiResponse(responseCode = "200", description = "List of bookings", content = {
                            @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Object.class)))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = BookingsApi.PATH_BOOKINGS_TENANTS_TENANT_ID_GET,
            produces = {"application/json"}
    )
    default ResponseEntity<Page<BookingDto>> bookingsTenantsTenantIdGet(
            Pageable page, @PathVariable Long tenantId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ null, null ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

    String PATH_BOOKINGS_TENANTS_TENANT_ID_GET = "/bookings/tenants/{tenantId}";

}