# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Atlas Sports API
  version: 1.0.0
servers:
- url: http://localhost:8080/api
...
paths:
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
          description: Successful login
      summary: Login user
      tags:
      - Auth
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Auth
  /auth/register:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
          description: Successful register
      summary: Register user
      tags:
      - Auth
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Auth
...
  /bookings:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/BookingDto"
                type: array
          description: List of bookings
      summary: Get all bookings (Admin can filter by tenant)
      tags:
      - Bookings
      x-accepts:
      - application/json
      x-tags:
      - tag: Bookings
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertBookingDto"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDto"
          description: Booking created
      summary: Create a booking
      tags:
      - Bookings
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Bookings
  /bookings/tenants/{tenantId}:
    get:
      parameters:
      - explode: false
        in: path
        name: tenantId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/BookingDto"
                type: array
          description: List of bookings
      summary: Get all bookings for tenant
      tags:
      - Bookings
      x-accepts:
      - application/json
      x-tags:
      - tag: Bookings
  /bookings/me/upcoming:
    get:
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: sort
        required: false
        schema:
          example: "createdAt,desc"
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_bookings_me_upcoming_get_200_response"
          description: List of upcoming bookings
      summary: Get upcoming bookings for current user
      tags:
      - Bookings
      x-accepts:
      - application/json
      x-tags:
      - tag: Bookings
  /bookings/me/history:
    get:
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      - explode: true
        in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: sort
        required: false
        schema:
          example: "createdAt,desc"
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_bookings_me_upcoming_get_200_response"
          description: Paginated list of past bookings
      summary: Get past bookings for current user
      tags:
      - Bookings
      x-accepts:
      - application/json
      x-tags:
      - tag: Bookings
...
components:
...
  schemas:
    LoginRequest:
      type: object
    LoginResponse:
      type: object
    RegisterRequest:
      type: object
    RegisterResponse:
      type: object
...
    BookingDto:
      type: object
    UpsertBookingDto:
      type: object
...
    PageResponse:
      properties:
        content:
          items:
            type: object
          type: array
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
      type: object
...
    _bookings_me_upcoming_get_200_response:
      allOf:
      - $ref: "#/components/schemas/PageResponse"
      - properties:
          content:
            items:
              $ref: "#/components/schemas/BookingDto"
            type: array
        type: object
      example:
        size: 6
        last: true
        totalPages: 5
        page: 0
        content:
        - null
        - null
        first: true
        totalElements: 1